account, apikey, contact modules added with basic APIs in new archite…

+++ db/migrations/000002_create_api_key_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE api_keys (
   id                  SERIAL PRIMARY KEY,
-  account_id          INTEGER NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
+  account_id          BIGINT NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
   key                 TEXT NOT NULL UNIQUE,
   created_at          TIMESTAMP DEFAULT now(),
   updated_at          TIMESTAMP DEFAULT now(),
